RemoveNaNColumns: Remove to preprocess
