+++ Orange/preprocess/score.py
@@ -8,8 +8,7 @@ from sklearn import feature_selection as skl_fss
 from Orange.data import Domain, Variable, DiscreteVariable, ContinuousVariable
 from Orange.data.filter import HasClass
 from Orange.misc.wrapper_meta import WrapperMeta
-from Orange.preprocess.fss import RemoveNaNColumns
-from Orange.preprocess.preprocess import Discretize, SklImpute
+from Orange.preprocess.preprocess import Discretize, SklImpute, RemoveNaNColumns
 from Orange.preprocess.util import _RefuseDataInConstructor
 from Orange.statistics import contingency, distribution
 from Orange.util import Reprable
